2023-08-01 End of the day
• Fixed ul generation in menu.html
• Changed indentation in most files
• Changed price from int to BigDecimal
• Added one state for OrderItems in OrderItemsStatus

+++ pizza12/src/main/java/fr/eni/pizza12/dal/ProductRepositoryImpl.java
@@ -39,7 +39,7 @@ public class ProductRepositoryImpl implements ProductRepository {
       ProductEntity productEntity = new ProductEntity();
       productEntity.setProductId(rs.getInt("productId"));
       productEntity.setProductName(rs.getString("productName"));
-      productEntity.setProductPrice(rs.getInt("productPrice"));
+      productEntity.setProductPrice(rs.getBigDecimal("productPrice"));
       productEntity.setProductActive(rs.getBoolean("isActive"));
       productEntity.setCategory(categoryRepositoryImpl.getCategoryById(rs.getInt("categoryId")));
       return productEntity;
@@ -98,7 +98,7 @@ public class ProductRepositoryImpl implements ProductRepository {
       public void setValues(PreparedStatement preparedStatement) throws SQLException {
         preparedStatement.setInt(1, productEntity.getProductId());
         preparedStatement.setString(2, productEntity.getProductName());
-        preparedStatement.setInt(3, productEntity.getProductPrice());
+        preparedStatement.setBigDecimal(3, productEntity.getProductPrice());
         preparedStatement.setBoolean(4, productEntity.isProductActive());
         preparedStatement.setInt(5, productEntity.getCategory().getCategoryId());
         preparedStatement.setInt(6, productEntity.getProductType().getProductTypeId());
@@ -147,7 +147,7 @@ public class ProductRepositoryImpl implements ProductRepository {
       public void setValues(PreparedStatement preparedStatement) throws SQLException {
         preparedStatement.setInt(1, productEntity.getProductId());
         preparedStatement.setString(2, productEntity.getProductName());
-        preparedStatement.setInt(3, productEntity.getProductPrice());
+        preparedStatement.setBigDecimal(3, productEntity.getProductPrice());
         preparedStatement.setBoolean(4, productEntity.getProductActive());
         preparedStatement.setInt(5, productEntity.getCategory().getCategoryId());
         preparedStatement.setInt(6, productEntity.getProductType().getProductTypeId());
